Don't use dynamic libraries everywhere.
In a branch I've recently hit this error:
```
error: cannot satisfy dependencies so `libc` only shows up once
  |
  = help: having upstream crates all available in one format will likely make this go away
```

What the error means to say is that two unrelated crates in our
workspace depend on the `libc` crate, and this would cause code
duplication and ultimately symbol clashing at link time (had the
compiler not caught this earlier).

One solution is to make one crate "responsible" for importing libc, and
then other crates would import it via a mere `extern crate`. Whilst
this would work, it would contort the repo into an odd structure, and
further, we may have to do this dance later for other deps other than
`libc`.

A more palatable solution is to use a dedicated `cdylib` crate for
exporting to C and then all other crates are regular `rlibs`. This way
there can be no duplicate deps.

This change therefore introduces `ykcapi` as our sole `cdylib` and
reverts all other workspace members to the default crate type. We also
remove our cargo config which gave preference to dynamic libs.

Thank you Bjorn and Jonas for the Zulip discussion on this. I love that
the Rust community is open to these kinds of informal discussions.
