--- .cargo/config.toml
@@ -1,2 +1,0 @@
-[build]
-rustflags = ["-Cprefer-dynamic"]
