+++ c_tests/tests/blockmap.c
@@ -3,20 +3,16 @@
 
 // Check the blockmap for this test program contains blocks.
 
+#include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <assert.h>
-
-// FIXME find a way to auto-generate these protos.
-void *yktrace_hwt_mapper_blockmap_new(void);
-size_t yktrace_hwt_mapper_blockmap_len(void *mapper);
-void yktrace_hwt_mapper_blockmap_free(void *mapper);
+#include <yk_testing.h>
 
 int
 main(int argc, char **argv)
 {
-    void *bm = yktrace_hwt_mapper_blockmap_new();
-    assert(yktrace_hwt_mapper_blockmap_len(bm) > 0);
-    yktrace_hwt_mapper_blockmap_free(bm);
+    void *bm = __yktrace_hwt_mapper_blockmap_new();
+    assert(__yktrace_hwt_mapper_blockmap_len(bm) > 0);
+    __yktrace_hwt_mapper_blockmap_free(bm);
     return (EXIT_SUCCESS);
 }
