+++ ykcapi/yk_testing.h
@@ -1,0 +1,5 @@
+// Functions exported only for testing.
+
+void *__yktrace_hwt_mapper_blockmap_new(void);
+size_t __yktrace_hwt_mapper_blockmap_len(void *mapper);
+void __yktrace_hwt_mapper_blockmap_free(void *mapper);
