+++ ykrt/Cargo.toml
@@ -5,9 +5,6 @@ authors = ["The Yorick Developers"]
 edition = "2018"
 license = "Apache-2.0 OR MIT"
 
-[lib]
-crate-type = ["dylib"]
-
 [dependencies]
 num_cpus = "1"
 once_cell = "1"
