+++ ykrt/src/lib.rs
@@ -2,9 +2,8 @@
 
 #![cfg_attr(test, feature(test))]
 
-mod capi;
 mod location;
 pub(crate) mod mt;
 
 pub use self::location::Location;
-pub use self::mt::MT;
+pub use self::mt::{HotThreshold, MT};
