+++ yktrace/Cargo.toml
@@ -5,9 +5,6 @@ authors = ["The Yorick Developers"]
 edition = "2018"
 license = "Apache-2.0 OR MIT"
 
-[lib]
-crate-type = ["dylib"]
-
 [dependencies]
 byteorder = "1.4.3"
 fxhash = "0.2.1"
